refactor: rewrite /karma to use collectors
Instead of having a separate command called karmaExplain, this
commit will make the karma and viewkarma commands use a
collector, which is an ad-hoc event listener that can listen
for followup interactions in a specific channel.
