+++ src/interactions/buttons/karma.ts
@@ -1,24 +1,20 @@
-import { ButtonInteraction, MessageActionRow } from "discord.js";
+import { ButtonInteraction } from "discord.js";
 import { ButtonInteractionHandler } from "../../lib/interaction";
-import { createKarmaToast } from "../../lib/karma";
-import Member from "../../lib/member";
-import { getExplainButton } from "./karmaExplain";
+import { handleKarmaInteraction } from "../../lib/karma/interaction";
+import { createToast } from "../../lib/response";
 
 export default class KarmaButton implements ButtonInteractionHandler {
   name = "karma_button";
 
   async handle(event: ButtonInteraction): Promise<void> {
-    const guildMember = await event.guild.members.fetch(event.user.id);
-    const member = new Member(guildMember);
-    const toast = await createKarmaToast(member, member.moderator);
-    await event.reply({
-      embeds: [toast],
-      components: [
-        new MessageActionRow({
-          components: [getExplainButton()],
-        }),
-      ],
-      ephemeral: true,
+    if (event.inGuild()) {
+      return handleKarmaInteraction(event, event.user.id);
+    }
+    const toast = createToast({
+      title: "Comando no apto para DM",
+      description: "Este comando sólo se puede usar en una guild",
+      severity: "error",
     });
+    return event.reply({ embeds: [toast] });
   }
 }
