+++ src/interactions/commands/karma.ts
@@ -1,27 +1,20 @@
-import { CommandInteraction, MessageActionRow } from "discord.js";
+import { CommandInteraction } from "discord.js";
 import { CommandInteractionHandler } from "../../lib/interaction";
-import { createKarmaToast } from "../../lib/karma";
-import Server from "../../lib/server";
-import { getExplainButton } from "../buttons/karmaExplain";
+import { handleKarmaInteraction } from "../../lib/karma/interaction";
+import { createToast } from "../../lib/response";
 
 export default class KarmaCommand implements CommandInteractionHandler {
   name = "karma";
 
-  async handle(command: CommandInteraction): Promise<void> {
-    if (command.inGuild()) {
-      await command.deferReply({ ephemeral: true });
-
-      const server = new Server(command.guild);
-      const member = await server.member(command.user);
-      const toast = await createKarmaToast(member, member.moderator);
-      await command.editReply({
-        embeds: [toast],
-        components: [
-          new MessageActionRow({
-            components: [getExplainButton()],
-          }),
-        ],
-      });
+  async handle(event: CommandInteraction): Promise<void> {
+    if (event.inGuild()) {
+      return handleKarmaInteraction(event, event.user.id);
     }
+    const toast = createToast({
+      title: "Comando no apto para DM",
+      description: "Este comando sólo se puede usar en una guild",
+      severity: "error",
+    });
+    return event.reply({ embeds: [toast] });
   }
 }
